Changed to use 'git mv' instead of 'mv'.
Added modified version of Rename: Rename2, which does a 'git mv' instead
of 'mv'.  module-rename: Changed to use Rename2 instead of Rename and
added a BEGIN block to load the local path of Rename2 based on the local
path of module-rename - so that the 'git mv' extension, which is not in
the official version, can be used locally.