Back to 'mv' as default and 'git mv' with -g option
Changed module-rename and Rename2.pm back, essentially, to implement the
default behavior of using (filesystem) mv instead of 'git mv'; and added
a -g option to module-rename, with corresponding logic in Rename2.pm to
specify use of 'git mv' instead of 'mv'.  Tested on a small repository
and verified that the default - mv - behavior is the same as the
original module-rename script/library and that the -g (git mv) behavior
is still correct.  Also, a couple small "improvements" (e.g.,
'use Modern::Perl').

Added .gitignore file.